Telecomm: Set MSIM audio params using the proper ID
It is assumed that the ID of every PhoneAccountHandle is the string
equivalent of an int, however, the ID can also includes hex chars or,
in case of the emergency account handle, it's just "E". This causes
issues when we try to get the sub ID using Integer.parseInt(),
so don't do it.

Change-Id: Iae6cfacaf65bb2791b6fa9ae491bb4ad6edbe936
(cherry picked from commit c2ed09b59edf95548299f1219f76c4da4f85d75f)